+++ src/kernels/bwd_bias.cl
@@ -16,7 +16,7 @@
 
 __kernel 
 __attribute__((reqd_work_group_size(WG_SIZE,1,1)))
-void bwd_bias(int features,int over,__global dtype *dy,int dy_offset,__global dtype *dx,int dx_offset,int dx_stride,float beta)
+void bwd_bias(int features,int over,__global dtype *dy,ulong dy_offset,__global dtype *dx,ulong dx_offset,int dx_stride,float beta)
 {
     dy += dy_offset;
     dx += dx_offset;
